% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_cos.R
\name{fn_cos}
\alias{fn_cos}
\title{Cosine Function}
\usage{
fn_cos(x, k)
}
\arguments{
\item{x}{the number (or vector of numbers) to be transformed}

\item{k}{the number of terms to be used in the series expansion beyond the constant 1. The value of k is always greater than or equal to 1.}
}
\value{
Returns the cosine using truncated series expansion.
}
\description{
Compute the cosine of a number
}
\details{
This function computes the cosine of a number using the truncated series expansion.
}
\examples{
fn_cos(4,9)

}
