% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_sin.R
\name{fn_sin}
\alias{fn_sin}
\title{Sine Function}
\usage{
fn_sin(x, k)
}
\arguments{
\item{x}{the number (or vector of numbers) to be transformed}

\item{k}{the number of terms to be used in the series expansion beyond the constant 1. The value of k is always greater than or equal to 1.}
}
\value{
Returns the sine using truncated series expansion.
}
\description{
Compute the sine of a number
}
\details{
This function computes the sine of a number using the truncated series expansion.
}
\examples{
fn_sin(6,15)

}
