% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3_ci_class_constructor.R
\name{make_ci_class}
\alias{make_ci_class}
\title{ci_class constructor}
\usage{
make_ci_class(x)
}
\arguments{
\item{x}{vector containing numeric values of ci_class}
}
\value{
none
}
\description{
Creates a ci_class object
}
\details{
This function transforms an input to an object of ci_class. The input must be a numeric vector. This function is useful because the calculate_CI function in this package only takes input of ci_class.
}
\examples{
x = c(0,1,2,3)
x = make_ci_class(x)
class(x)
}
