% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3_ci_class_print_method.R
\name{print.ci_class}
\alias{print.ci_class}
\title{Print Method for ci_class}
\usage{
\method{print}{ci_class}(x)
}
\arguments{
\item{x}{a ci_class object}
}
\value{
Returns the ci_class object
}
\description{
Prints the number of observations in a ci_class object
}
\details{
This function is ci_class method and it prints the number of observations in a ci_class object. It is invisibly returns the ci_class object.
}
\examples{
x = c(0,1,2,3)
x = make_ci_class(x)
print(x)

}
