% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Calculate_CI.R
\name{calculate_CI}
\alias{calculate_CI}
\title{Calculate Confidence Intervals}
\usage{
calculate_CI(x, conf = 0.95)
}
\arguments{
\item{x}{vector containing numeric values}

\item{conf}{confidence level used to generate the confidence interval}
}
\value{
Returns lower- and upperbound of the confidence interval
}
\description{
Calculate the confidence interval around the mean of a vector
}
\details{
This function calculates and returns the lower- and upperbounds of the confidence interval about the mean of a vector. The vector and the desired confidence level are passed as arguments to the function. The default level of confidence is 0.95.
}
\examples{
set.seed(1234)
x = sample(10:1000,20)
calculate_CI(x,conf=0.95)

}
